move graph-types to core
