+++ test/types/graph-types.test.ts
@@ -18,6 +18,8 @@
  */
 
 import {
+  int,
+  Integer,
   Node,
   Path,
   PathSegment,
@@ -28,8 +30,7 @@ import {
   isPathSegment,
   isRelationship,
   isUnboundRelationship
-} from '../../types/graph-types'
-import { int, Integer } from 'neo4j-driver-core'
+} from 'neo4j-driver-core'
 
 const node1: Node = new Node(int(1), ['Person', 'Employee'], { name: 'Alice' })
 const node1String: string = node1.toString()
