+++ types/result-summary.d.ts
@@ -17,8 +17,7 @@
  * limitations under the License.
  */
 
-import { Integer } from 'neo4j-driver-core'
-import { NumberOrInteger } from './graph-types'
+import { Integer, NumberOrInteger } from 'neo4j-driver-core'
 
 declare interface ResultSummary<T extends NumberOrInteger = Integer> {
   query: { text: string; parameters: { [key: string]: any } }
