+++ types/session.d.ts
@@ -20,7 +20,7 @@
 import Transaction from './transaction'
 import QueryRunner, { Parameters } from './query-runner'
 import Result from './result'
-import { NumberOrInteger } from './graph-types'
+import { NumberOrInteger } from 'neo4j-driver-core'
 
 declare type TransactionWork<T> = (tx: Transaction) => T | Promise<T>
